feat: enhance face detection and user feedback
Improved face detection in the questionnaire and bloodshot eye percentage tests by adding a warning message if no face is detected for
 a certain number of trials. This ensures that the user is properly positioned before capturing an image.

Added a warning dialog if the detected user is not the enrolled user for a certain number of trials in the face camera and questionnaire fragments. This prevents cheating and ensures that the correct user is taking the test.

Updated the
 navigation flow to return to the user profile screen if no face is detected in the bloodshot eye percentage test. This allows the user to retake the test.

Removed the result and advice section from the questionnaire interaction guidelines.

Added a completed dialog to the questionnaire fragment that is displayed when the questionnaire is finished. This informs the user that the questionnaire is complete and navigates them back to the user profile screen.

Added a warning overlay view and layout to the questionnaire fragment to display warning messages to the user. This provides a clear visual indication of the warning.

Updated the face camera fragment to handle exceptions while parsing image files. This prevents
 the app from crashing if there is an error with the image files.