feat: add remote config to check for updates
Added Firebase Remote Config to check for new app updates.

If a new update is available, a dialog will
 be displayed prompting the user to update.
The dialog includes a button that will take the user to the app's update URL.

Added a new Utils object with helper functions to get the app version and version code.
Added a new ApplicationMain class to initialize the app version and version code.
Updated the
 MainActivity to handle the remote config and display the update dialog.
Added new strings for the update dialog title, message, and button text.
Added the Firebase Config dependency to the app's build.gradle file.
Updated the AndroidManifest.xml file to include the ApplicationMain class.