feat: add glide and update permissions
Added Glide library for efficient image loading and caching.

Updated permissions to include `ACCESS_MEDIA_LOCATION` for Android
 versions above 32 and `READ_EXTERNAL_STORAGE` and `WRITE_EXTERNAL_STORAGE` for versions below 33.

Modified the `downloadImageFromFirebase` function to use Glide for downloading and saving images to external storage.

Updated the `hasPermissions` function to check for the new permissions
.

Changed the navigation flow to navigate to the start screen after permissions are granted.

Updated the `userDataItem` data class to use lowercase `id`.