feat: add user test history and start new test
Added functionality to display the user's last test date and state on the user profile screen.
Added
 a "Re-Test" button to the user profile screen to start a new test.
Added a dialog to inform the user about the test requirements before starting a new test.
Added a new test entry to the database when the user starts a new test.
Updated the navigation flow to navigate to the face camera
 fragment after starting a new test.
Updated the UI to display the last test state instead of the last test score.
Added a new model class for user tests.
Added utility functions to convert data to JSON array, get the current date, and get the latest test from a list of tests.
Updated the start screen fragment to set the current user ID when a new user is created or an existing user is selected.